test junit adresses

+++ src/main/java/fr/miage/mesures/Quantite.java
@@ -11,7 +11,7 @@ package fr.miage.mesures;
  * @author E12B336Z
  */
 public class Quantite {
-	private int myValeur;
+	private double myValeur;
         private Unite unite;
 
     public Quantite(int myValeur, Unite unite) {
@@ -27,7 +27,7 @@ public class Quantite {
         this.unite = unite;
     }
         
-    public int getMyValeur() {
+    public double getMyValeur() {
         return myValeur;
     }
